Make the following changes:
,

Find this element:

  <div class="p-6 pt-0">

with the text content:

    Video URLValidate & Process Link

Relevant files:

And change it as follows:

I'd like to implement a feature that allows users to download videos from YouTube. Please design an interface where users can paste a YouTube link and then initiate the download process. Ensure the design accommodates this functionality clearly and intuitively. Thank you.