Make the following changes:
,

Find this element:

  <Button size="lg" className="h-12 px-6 sm:h-14 sm:px-8 text-base sm:text-lg group/button self-start mt-2 animate-fade-in-up animation-delay-400">

with the text content:

    Watch Trailer

Relevant files:

    - https://6000-firebase-studio-1746796893098.cluster-oayqgyglpfgseqclbygurw4xd4.cloudworkstations.dev/_next/static/chunks/src_cac2839c._.js
    - https://6000-firebase-studio-1746796893098.cluster-oayqgyglpfgseqclbygurw4xd4.cloudworkstations.dev/_next/static/chunks/node_modules_a31e2e15._.js
    - https://6000-firebase-studio-1746796893098.cluster-oayqgyglpfgseqclbygurw4xd4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1746796893098.cluster-oayqgyglpfgseqclbygurw4xd4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

You have changed the language of my application to Swahili, and everything was successfully translated—except the 'Watch Trailer' button, which is still in English. I would like this button to be translated into Swahili as well to keep the app fully consistent. Please update it accordingly. Suggested translation: "Tazama Trela".